Refactoring Добавил удаление рецептов

+++ src/main/java/com/miti/server/repository/IngredientContextRepository.java
@@ -11,8 +11,8 @@ import java.util.List;
 
 @Repository
 public interface IngredientContextRepository extends JpaRepository<IngredientContext, Long> {
-  List<IngredientContext> getIngredientContextsByAmountLessThan(Integer amount);
-  List<IngredientContext> getIngredientContextByAmountGreaterThan(Integer amount);
+  List<IngredientContext> getIngredientContextsByAmountLessThan(Long amount);
+  List<IngredientContext> getIngredientContextByAmountGreaterThan(Long amount);
   List<IngredientContext> getIngredientContextsByMeasure(Measure measure);
   List<IngredientContext> getIngredientContextsByIngredient(Ingredient ingredient);
   List<IngredientContext> getIngredientContextsByRecipeIngredients(Recipe recipe);
